run-clean.sh: fix the return non-zero

+++ run-clean.sh
@@ -1,5 +1,13 @@
 #!/bin/bash -eu
+if [ $# -lt 1 ];then
+    echo "Usage: $0 <command> [Command Args] ... " > /dev/stderr
+    exit 1
+fi
 Cmd=$1
 shift
 $Cmd ${@-}
-pkill -P $$
+if pgrep -P $$ -f $Cmd ${@-} ;then
+    ## Only kill children if they exist, otherwise pkill return non zero
+    pkill -P $$
+fi
+
